Break round order into a custom hook
